Fixed FCFS calculations

+++ src/process_scheduling.c
@@ -36,70 +36,70 @@ void virtual_cpu(ProcessControlBlock_t *process_control_block)
 bool first_come_first_serve(dyn_array_t *ready_queue, ScheduleResult_t *result) 
 {
 	//check that input is valid and contains PCB's'
-	if(ready_queue == NULL || result == NULL || ready_queue->data_size != sizeof(ProcessControlBlock_t))
-	{
-		return false;
-	}
-	
-	uint32_t clockTime = 0;
-	uint32_t busyTime = 0;
-	uint32_t waitTime = 0;
-
-	bool fin = false;
-	while(!fin)
-	{
-		//initialize a current PCB and an iterator
-		ProcessControlBlock_t* cur = NULL;
-		uint8_t* iter = (uint8_t *) ready_queue->array;
-		
-		fin = true;
-
-		//store next shortest and valid job in cur and update fin;
-		for (size_t idx = 0; idx < ready_queue->size; ++idx, iter += ready_queue->data_size) 
-		{
-			ProcessControlBlock_t* walker = (ProcessControlBlock_t*) iter;
-			
-			//check if current walker is valid for comparison, updating fin or returning false if encountering invalid process
-			if(walker->remaining_burst_time > 0)
-				fin = false;
-			else if(!(walker->started))
-				return false;
-			else
-				continue;
-
-			//update current PCB if walker has arrived and has more recent arrival (or if the cur is NULL, so we can initialize a PCB)
-			if((cur == NULL || walker->arrival < cur->arrival) && walker->arrival <= clockTime)
-				cur = walker;
-		}
-		
-		//if you can't find anything valid, increment time, otherwise operate on the PCB
-		if(cur == NULL)
-		{
-			clockTime++;
-			busyTime++;
-		}
-		else
-		{
-			waitTime += (clockTime - cur->arrival);
-			cur->started = true;
-			while(cur->remaining_burst_time > 0)
-			{
-				virtual_cpu(cur);
-				clockTime++;
-			}
-		}
-	}
-
-	//decrement time once to account for the final wait after finishing
-	clockTime--;
-	busyTime--;
-
-	//calculate the result values
-	result->average_waiting_time = (float)waitTime / ready_queue->size;
-	result->total_run_time = (unsigned long) clockTime - busyTime;
-	result->average_turnaround_time = (float) (result->average_waiting_time + result->total_run_time) / ready_queue->size;
-	
-	return true;
+    if(ready_queue == NULL || result == NULL || ready_queue->data_size != sizeof(ProcessControlBlock_t))
+    {
+    	return false;
+    }
+    
+    uint32_t clockTime = 0;
+    uint32_t busyTime = 0;
+    uint32_t waitTime = 0;
+    
+    bool fin = false;
+    while(!fin)
+    {
+    	//initialize a current PCB and an iterator
+    	ProcessControlBlock_t* cur = NULL;
+    	uint8_t* iter = (uint8_t *) ready_queue->array;
+    	
+    	fin = true;
+    
+    	//store next shortest and valid job in cur and update fin;
+    	for (size_t idx = 0; idx < ready_queue->size; ++idx, iter += ready_queue->data_size) 
+    	{
+    		ProcessControlBlock_t* walker = (ProcessControlBlock_t*) iter;
+    		
+    		//check if current walker is valid for comparison, updating fin or returning false if encountering invalid process
+    		if(walker->remaining_burst_time > 0)
+    			fin = false;
+    		else if(!(walker->started))
+    			return false;
+    		else
+    			continue;
+    
+    		//update current PCB if walker has arrived and has more recent arrival (or if the cur is NULL, so we can initialize a PCB)
+    		if((cur == NULL || walker->arrival < cur->arrival) && walker->arrival <= clockTime)
+    			cur = walker;
+    	}
+    	
+    	//if you can't find anything valid, increment time, otherwise operate on the PCB
+    	if(cur == NULL)
+    	{
+    		clockTime++;
+    		busyTime++;
+    	}
+    	else
+    	{
+    		waitTime += (clockTime - cur->arrival);
+    		cur->started = true;
+    		while(cur->remaining_burst_time > 0)
+    		{
+    			virtual_cpu(cur);
+    			clockTime++;
+    		}
+    	}
+    }
+    
+    //decrement time once to account for the final wait after finishing
+    clockTime--;
+    busyTime--;
+    
+    //calculate the result values
+    result->average_waiting_time = (float)waitTime / ready_queue->size;
+    result->total_run_time = (unsigned long) clockTime - busyTime;
+    result->average_turnaround_time = (float) result->average_waiting_time + ((float) result->total_run_time / ready_queue->size);
+    
+    return true;
 }
 
 //this function will schedule the tasks based on which has the shortest job, and it does NOT exhibit preemptive behavior
